+++ accounts/models.py
@@ -260,17 +260,3 @@ def create_user_profile(sender, instance, created, **kwargs):
         profile.save()
 
 post_save.connect(create_user_profile, sender=User)
-
-class DonationCampaign(models.Model):
-    goal = models.DecimalField(max_digits=6, decimal_places=2)
-    date_start = models.DateField()
-
-
-class Donation(models.Model):
-    email = models.CharField(max_length=255)
-    display_name = models.CharField(max_length=255)
-    amount = models.DecimalField(max_digits=5, decimal_places=2)
-    transaction_id = models.CharField(max_length=255, blank=True)
-    currency = models.CharField(max_length=100) # Should always be EUR
-    created = models.DateTimeField(auto_now_add=True)
-    campaign = models.ForeignKey(DonationCampaign, null=True)
