+++ donations/urls.py
@@ -1,0 +1,31 @@
+# -*- coding: utf-8 -*-
+
+#
+# Freesound is (c) MUSIC TECHNOLOGY GROUP, UNIVERSITAT POMPEU FABRA
+#
+# Freesound is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+#
+# Freesound is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+#
+# Authors:
+#     See AUTHORS file.
+#
+
+from django.conf.urls import patterns, url, include
+import views
+
+urlpatterns = [
+    url(r'^donate/$', views.donate, name="donate"),
+    url(r'^donors/$', views.DonationsList.as_view(), name="donors"),
+    url(r'^donation-complete/$', views.donation_complete, name="donation-complete"),
+
+]
