+++ general/management/commands/create_front_page_caches.py
@@ -23,7 +23,7 @@ from django.conf import settings
 from django.template.loader import render_to_string
 from django.core.cache import cache
 from django.db.models import Sum
-import accounts.models
+import donations.models
 import logging
 logger = logging.getLogger("web")
 
@@ -39,11 +39,12 @@ class Command(NoArgsCommand):
         rss_cache = render_to_string('rss_cache.html', locals())
         cache.set("rss_cache", rss_cache, 2592000) # 30 days cache
 
-        campaign = accounts.models.DonationCampaign.objects.order_by('date_start').last()
-        donations = accounts.models.Donation.objects\
+        campaign = donations.models.DonationCampaign.objects.order_by('date_start').last()
+        all_donations = donations.models.Donation.objects\
                 .filter(campaign=campaign).all().aggregate(Sum('amount'))
-        donations_goal = campaign.goal
-        params = {'remains': int(donations_goal - (donations['amount__sum'] or 0)),
-                  'percent_towards_goal': int((donations['amount__sum'] or 0) / donations_goal * 100)}
-        donations_cache = render_to_string('donations_cache.html', params)
-        cache.set("donations_cache", donations_cache, 2592000)  # 30 days cache
+        if campaign:
+            donations_goal = campaign.goal
+            params = {'remains': int(donations_goal - (all_donations['amount__sum'] or 0)),
+                      'percent_towards_goal': int((all_donations['amount__sum'] or 0) / donations_goal * 100)}
+            donations_cache = render_to_string('donations_cache.html', params)
+            cache.set("donations_cache", donations_cache, 2592000)  # 30 days cache
